refactor: converter valores monetários de double para BigDecimal

+++ src/main/java/com/marafigo/finance/entities/dto/UserBalanceDTO.java
@@ -6,10 +6,12 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
+
 @Getter @Setter
 @NoArgsConstructor @AllArgsConstructor
 public class UserBalanceDTO implements Serializable {
-    private Double totalIncome;
-    private Double totalExpense;
-    private Double balance;
+    private BigDecimal totalIncome;
+    private BigDecimal totalExpense;
+    private BigDecimal balance;
 }
